Fixed minor warning when compiling on windows (type matches) and minor clean ups

+++ src/Correlate.cpp
@@ -213,10 +213,10 @@ void Correlator::get_x_axis(unsigned long long** x_axis, int* n_out){
 }
 
 
-void Correlator::make_fine(unsigned int* tac_1, int n_tac_1, unsigned int* tac_2, int n_tac_2, int n_tac){
+void Correlator::make_fine(unsigned int* tac_1, unsigned int n_tac_1, unsigned int* tac_2, int unsigned n_tac_2, unsigned int n_tac){
     if(!is_fine){
-        make_fine_times(t1, n_t1, tac_1, n_tac);
-        make_fine_times(t2, n_t2, tac_2, n_tac);
+        make_fine_times(t1, (unsigned int) n_t1, tac_1, n_tac);
+        make_fine_times(t2, (unsigned int) n_t2, tac_2, n_tac);
         is_fine = true;
     }
 }
